backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
asyncpg
redis
ollama
faiss-cpu
numpy
scikit-learn
python-dotenv
httpx
